requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.4
